data_collection/ThirdEye-II/requirements.txt
gymnasium==0.29.1
Flask==2.0.0
Flask-SocketIO==4.3.1
python-engineio==3.13.2
python-socketio==4.5.1
eventlet==0.35.1
pandas==2.2.3
Werkzeug==2.0.3
pillow==10.2.0
tqdm==4.66.4
pygame==2.5.2

# Additional stuff for DL
torchvision==0.18.0
torch==2.3.0
lightning==2.2.5
torchinfo==1.8.0
torchmetrics==1.3.2

# Plotting stuff
matplotlib
seaborn

# tensorflow==2.12.0
# opencv-python==4.9.0.80
# matplotlib==3.8.4
